Python/Flight Search Engine/requirements.txt
psycopg2-binary
pyshorteners